# Improve the performance of list catalog objects in a bucket request by reducing the redundant DB queries