--- a/src/main/java/org/ow2/proactive/catalog/repository/BucketGrantRepository.java
+++ b/src/main/java/org/ow2/proactive/catalog/repository/BucketGrantRepository.java
@@ -57,11 +57,18 @@ public interface BucketGrantRepository extends JpaRepository<BucketGrantEntity, 
     List<BucketGrantEntity> findAllAccessibleBucketGrantsAssignedToAUsernameInsideABucket(String username,
             long bucketId);
 
-    @Query(value = "SELECT bge FROM BucketGrantEntity bge WHERE bge.grantee in ?1 And bge.granteeType='group' AND bge.bucketEntity.id=?2 AND bge.accessType<>'noAccess'")
+    @Query(value = "SELECT bge FROM BucketGrantEntity bge WHERE bge.grantee in ?1 AND bge.bucketEntity.id=?2 And bge.granteeType='group' AND bge.accessType<>'noAccess'")
     List<BucketGrantEntity> findAllAccessibleBucketGrantsAssignedToTheUserGroupsInsideABucket(List<String> userGroup,
             long bucketId);
 
     @Query(value = "SELECT bge FROM BucketGrantEntity bge WHERE bge.accessType='noAccess'")
     List<BucketGrantEntity> findAllBucketGrantsWithNoAccessRight();
+
+    @Query(value = "SELECT bge FROM BucketGrantEntity bge WHERE bge.grantee = ?1 AND bge.bucketEntity.id=?2 And bge.granteeType='user' AND bge.accessType='noAccess'")
+    List<BucketGrantEntity> findBucketGrantsAssignedToAUsernameWithNoAccessRight(String username, long bucketId);
+
+    @Query(value = "SELECT bge FROM BucketGrantEntity bge WHERE bge.grantee in ?1 AND bge.bucketEntity.id=?2 And bge.granteeType='group' AND bge.accessType='noAccess'")
+    List<BucketGrantEntity> findBucketGrantsAssignedToUserGroupsWithNoAccessRight(List<String> userGroup,
+            long bucketId);
 
 }