--- a/src/main/java/org/ow2/proactive/catalog/repository/CatalogObjectGrantRepository.java
+++ b/src/main/java/org/ow2/proactive/catalog/repository/CatalogObjectGrantRepository.java
@@ -83,8 +83,14 @@ public interface CatalogObjectGrantRepository extends JpaRepository<CatalogObjec
             findAllObjectGrantsWithNoAccessRightsAndAssignedToAUserGroup(List<String> userGroups);
 
     @Query(value = "SELECT coge FROM CatalogObjectGrantEntity coge WHERE coge.grantee = ?1 AND coge.granteeType='user' AND coge.accessType<>'noAccess'")
-    List<CatalogObjectGrantEntity> findAllObjectGrantsAssignedToAUser(String username);
+    List<CatalogObjectGrantEntity> findAllAccessibleObjectGrantsAssignedToAUser(String username);
 
     @Query(value = "SELECT coge FROM CatalogObjectGrantEntity coge WHERE coge.grantee in ?1 AND coge.granteeType='group' AND coge.accessType<>'noAccess'")
+    List<CatalogObjectGrantEntity> findAllAccessibleObjectGrantsAssignedToUserGroups(List<String> userGroup);
+
+    @Query(value = "SELECT coge FROM CatalogObjectGrantEntity coge WHERE coge.grantee = ?1 AND coge.granteeType='user'")
+    List<CatalogObjectGrantEntity> findAllObjectGrantsAssignedToAUser(String username);
+
+    @Query(value = "SELECT coge FROM CatalogObjectGrantEntity coge WHERE coge.grantee in ?1 AND coge.granteeType='group'")
     List<CatalogObjectGrantEntity> findAllObjectGrantsAssignedToUserGroups(List<String> userGroup);
 }