… src/main/java/org/ow2/proactive/catalog/service/BucketGrantService.java
… src/main/java/org/ow2/proactive/catalog/service/CatalogObjectGrantService.java
… src/main/java/org/ow2/proactive/catalog/service/GrantAccessTypeHelperService.java
… src/main/java/org/ow2/proactive/catalog/service/GrantRightsService.java
… src/test/java/org/ow2/proactive/catalog/service/BucketGrantServiceTest.java
… src/test/java/org/ow2/proactive/catalog/service/GrantRightServiceTest.java