This PR aims to improve the performance of catalog requests of listing buckets and catalog objects:
- GET buckets/ 
- GET buckets/BUCKET_NAME/resources/

Performance test data on a relatively slow PC (local requests):
- GET buckets/: **before** 2000 ms - 3000 ms **after** 800 ms - 1200 ms
- GET buckets/BUCKET_NAME/resources/: **before** 5000 ms - 6000 ms **after** 400 ms - 800 ms 

Important changes:
- avoid redundant DB call: retrieve all the needed data at once (or with fewer DB queries), then filter it to get the needed data. At least, it should be avoided to send DB queries in a loop.
- reduce unnecessary loop
- use Set instead of List when a quick match needed

Small changes:
- reduce duplicate coding
- remove not used functions
- reduce double negative coding
- reduce embedded brackets levels

Potential bugs to be fixed in the future:
- The function `BucketGrantService.getTheNumberOfAccessibleObjectsInTheBucket` and `GrantRightsService.removeAllObjectsWithNoAccessGrant` need to compare grants priority. The current version may cause bugs when the user has multiple conflict group grants (or user grants) over the same catalog object / bucket.