@marounkoussaifi 2022-05-09
LGTM! Thank you @tao-xinxiu for the improvements. We need to run manual tests once this PR is merged to guarantee that the changes did not affect the normal behavior.